moving to gradle build system (#33)
* initial commit for moving to gradle

+++ oci-objectstorage-fixture/src/main/java/org/opensearch/fixtures/oci/NonJerseyServer.java
@@ -26,7 +26,7 @@ public class NonJerseyServer implements Closeable {
                 () -> {
                     try {
                         server.start();
-                        Thread.sleep(Integer.MAX_VALUE);
+                        Thread.sleep(Long.MAX_VALUE);
                     } catch (InterruptedException e) {
                         throw new RuntimeException(e);
                     }
